refactor actions and stores data for only ergo connector
