+++ app/ergo-connector/App.js
@@ -22,8 +22,8 @@ import { Routes } from './Routes';
 import { yoroiPolymorphTheme } from '../themes/PolymorphThemes';
 import { themeOverrides } from '../themes/overrides';
 import { translations } from '../i18n/translations';
-import type { StoresMap } from '../stores';
-import type { ActionsMap } from '../actions';
+import type { StoresMap } from './stores';
+import type { ActionsMap } from './actions';
 import { changeToplevelTheme } from '../themes';
 import ThemeManager from '../ThemeManager';
 import CrashPage from '../containers/CrashPage';
